Add if statemtn to tags.html
When an image was deleted that was the only image with a tag, the
tag had no images associate with it.  Therefor, when the tags.html
template would fire, it was looking for a link to the first image
associate with the tag.  Since there was no association, it threw
an out or range error.  Adding an if statement to check if there
was an image associated first, solved this problem.