Fix many to many delete issue
When deleting an image, deletion of the tags from the tag lookup
was trying to happen twice due to the cross relationship.  This was
causing sqlalchmy to try to delete something that wasn't there and
would cause an error.

I added a "viewonly=True" to the images backref to prevent it trying
to delete.  This appears to have fixed the issue.